+++ src/components/spartan/SDataTable/types.ts
@@ -1,9 +1,11 @@
 export type TDataTableProps = {
     cols: Record<string, string>[];
     pageSizes?: number[];
+    initialPageSize?: number;
     data: unknown[];
     sortable?: boolean;
     filtrable?: boolean;
     loading?: boolean;
+    pagination?: boolean;
     containerClass?: string;
 };